ProfileAction as an abstract, to support common profile actions merge,reconcile and sync
